rename Views components
